Update Community-Operators Publisher Workflow (#1307)
* Apply retry logic in check_defaults.sh

Apply retry logic in check_defaults.sh to prevent flakyness due
to conflicts by the validating webhook cause by side effects
from the previous changes.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>

* Update Community-Operators Publisher Workflow

Since we are now working with floating tags for hco-operator and hco-webhook in the CSV, there is no need to calculate their digests and re-commit to the CSV after building and pushing them.
Therefore, upon making a tag with a version, the publish action has been extended to:
1. Build and push the application images (hco-operator, hco-webhook).
2. Calculate and replace the floating tags of the application images with their digests, to be published on community-operators.
3. Build and push the bundle and index image in semver mode (remove 'replaces' field from the CSV).

Signed-off-by: orenc1 <ocohen@redhat.com>

* minor unrelated fixes

Signed-off-by: orenc1 <ocohen@redhat.com>

Co-authored-by: Simone Tiraboschi <stirabos@redhat.com>