WIP: split writing HyperConverged's status and body (#1409)
split writing HyperConverged's status and body (spec/metadata)

Since writing to the HyperConverged status must be a seperate write nad we can't write the whole custome resource as an automic operation, it caused several instbility and conflicts.

This PR split the right to two calls - if the code modified something in the body (should be with very low frequency, except for during upgrade), HCO will write only the body, and exit with requeue = true to force additional call, then we expect the body not to be modified again, letting HCO to write only the status this time.

This is WIP because the unit test can't fully check that in unit tests, as the client mock *does* write the status together with the body, so if there is a mistake here, it will not be found in unit tests, and we need the CI tests to validate.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>